Added initiation for some variables

+++ server/src/server.c
@@ -56,8 +56,8 @@ int initServer(Server *pServer) {
     pServer->payload.serverState = SERVER_WAITING;
     for (int i = 0; i < MAX_PLAYERS; i++) {
         pServer->payload.players[i].position = playerGetPosition(pServer->players[i]);
-        pServer->payload.players[i].sheetPosition.x = playerGetSheetPosition(pServer->players[i]).x;
-        pServer->payload.players[i].sheetPosition.y = playerGetSheetPosition(pServer->players[i]).y;
+        pServer->payload.players[i].sheetPosition.x = (float)playerGetSheetPosition(pServer->players[i]).x;
+        pServer->payload.players[i].sheetPosition.y = (float)playerGetSheetPosition(pServer->players[i]).y;
         pServer->payload.players[i].tick = 0;
     }
 
@@ -108,8 +108,8 @@ void handleTick(Server *pServer, ClientPayload payload, float const timestep) {
     if (!standingOnPlatform) { playerSetState(pPlayer, FALLING); }
 
     pServer->payload.players[payload.playerID].position = playerGetPosition(pPlayer);
-    pServer->payload.players[payload.playerID].sheetPosition.x = playerGetSheetPosition(pPlayer).x;
-    pServer->payload.players[payload.playerID].sheetPosition.y = playerGetSheetPosition(pPlayer).y;
+    pServer->payload.players[payload.playerID].sheetPosition.x = (float)playerGetSheetPosition(pPlayer).x;
+    pServer->payload.players[payload.playerID].sheetPosition.y = (float)playerGetSheetPosition(pPlayer).y;
     pServer->payload.players[payload.playerID].tick = payload.player.tick;
     return;
 }
